+++ golang/lib/client/do_nothing_client/do_nothing_client.go
@@ -20,17 +20,17 @@ func (client DoNothingClient) TrackCreateEnclave(enclaveId string) error {
 	return nil
 }
 
-func (client DoNothingClient) TrackStopEnclave() error {
+func (client DoNothingClient) TrackStopEnclave(enclaveId string) error {
 	logrus.Debugf("Do nothing client TrackStopEnclave called")
 	return nil
 }
 
-func (client DoNothingClient) TrackDestroyEnclave() error {
+func (client DoNothingClient) TrackDestroyEnclave(enclaveId string) error {
 	logrus.Debugf("Do nothing client TrackDestroyEnclave called")
 	return nil
 }
 
-func (client DoNothingClient) TrackCleanEnclave() error {
+func (client DoNothingClient) TrackCleanEnclave(shouldCleanAll bool) error {
 	logrus.Debugf("Do nothing client TrackCleanEnclave called")
 	return nil
 }